Extract Troll/Thief/Cyclops's IsDead+GodModeKill into a shared base class
Per review feedback on #375: the IsDead flag, the once-only guard, and the
ICanBeAttacked.GodModeKill wrapper were duplicated identically across all
three creatures. CombatGateCreatureBase (: ContainerBase, ICanBeAttacked)
now owns all of that; each subclass only implements OnDeath for its own
creature-specific removal effects.

Cyclops moves from ItemBase directly onto CombatGateCreatureBase (via
ContainerBase), since Troll and Thief already needed ContainerBase's Items
list and there's no way to share one base class across ItemBase- and
ContainerBase-rooted types otherwise. Verified this doesn't regress real
gameplay noun resolution ("kill the cyclops with the sword" still passes -
article-stripping happens in the intent parser, not in the container's
stricter HasMatchingNoun), and that the "put X in cyclops" container surface
this newly opens is the same untested-but-tolerated capability Troll and
Thief already had before this PR, not a new one unique to Cyclops.
