+++ ZorkOne/Item/Troll.cs
@@ -8,7 +8,7 @@ using ZorkOne.ActorInteraction;
 
 namespace ZorkOne.Item;
 
-public class Troll : ContainerBase, ICanBeExamined, ITurnBasedActor, ICanBeAttacked, ICanBeGivenThings, ICanHoldItems
+public class Troll : CombatGateCreatureBase, ICanBeExamined, ITurnBasedActor, ICanBeGivenThings, ICanHoldItems
 {
     private readonly GiveSomethingToSomeoneDecisionEngine<Troll> _giveHimSomethingEngine = new();
 
@@ -17,8 +17,6 @@ public class Troll : ContainerBase, ICanBeExamined, ITurnBasedActor, ICanBeAttac
     public bool IsUnconscious { get; set; }
 
     public bool IsStunned { get; set; }
-
-    public bool IsDead { get; set; }
 
     public IItem? ItemBeingHeld { get; set; }
 
@@ -110,19 +108,12 @@ public class Troll : ContainerBase, ICanBeExamined, ITurnBasedActor, ICanBeAttac
     }
 
     /// <summary>
-    /// Issue #374: shared death-state transition used by both a real fatal combat blow
-    /// (AdventurerVersusTrollCombatEngine.DeathBlow) and "god mode kill troll", so the two can never
-    /// drift out of sync. The axe always ends up in TrollRoom specifically - not wherever the troll's
-    /// own CurrentLocation happens to point - because a prior "god mode take troll" can move that
-    /// mutable state and would otherwise misdirect the drop.
+    /// Issue #374: the axe always ends up in TrollRoom specifically - not wherever the troll's own
+    /// CurrentLocation happens to point - because a prior "god mode take troll" can move that mutable
+    /// state and would otherwise misdirect the drop.
     /// </summary>
-    internal void Die()
+    protected override void OnDeath(IContext context)
     {
-        if (IsDead)
-            return;
-
-        IsDead = true;
-
         if (ItemBeingHeld is not null)
         {
             Repository.GetLocation<TrollRoom>().ItemPlacedHere(ItemBeingHeld);
@@ -133,12 +124,6 @@ public class Troll : ContainerBase, ICanBeExamined, ITurnBasedActor, ICanBeAttac
         // normal case, but also correctly handles the (god-mode-only) edge case where he was
         // previously moved into the player's inventory via "god mode take troll".
         Repository.DestroyItem(this);
-    }
-
-    public bool GodModeKill(IContext context)
-    {
-        Die();
-        return true;
     }
 }
